requirements.txt
# Core dependencies
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0

# Database
pymongo>=4.6.0

# Machine Learning
scikit-learn>=1.4.0
mlflow>=2.10.0

# Data Visualization
plotly>=5.18.0

# API and Web
requests>=2.31.0

# Azure Integration (for future deployment)
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Health Agent
torch>=2.2.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.15
openai>=1.12.0 